+++ lib/pages/home.dart
@@ -31,26 +31,24 @@ class _HomeState extends State<Home> {
   late List<New> news;
 
   Future<void> getData(String? category) async {
-
     cards = await service.getProducts(category, null);
     news = await service.getNews();
   }
 
-  Future<void> storeProductCart(Product product) async {
-    await service.storeProductCart(product);
-    // refreshPage();
+  Future<void> changeProductCart(Product product) async {
+    if (product.added) {
+      await service.deleteProductCart(product);
+    } else {
+      await service.storeProductCart(product);
+    }
+    refreshPage();
   }
 
-  Future<void> removeProductCart(Product product) async {
-    await service.deleteProductCart(product);
-    // refreshPage();
+  void refreshPage() {
+    setState(() {
+      _future = getData(categories[selectedCategory]['code']);
+    });
   }
-
-  // void refreshPage() {
-  //   setState(() {
-  //     _future = getData(categories[selectedCategory]['code']);
-  //   });
-  // }
 
   @override
   void initState() {
@@ -142,7 +140,7 @@ class _HomeState extends State<Home> {
                         child: Padding(
                           padding: const EdgeInsets.symmetric(vertical: 8.0),
                           child: CardOrder(
-                            buttonFunc: cards[index].added ? removeProductCart(cards[index]) : storeProductCart(cards[index]),
+                            buttonFunc: () => changeProductCart(cards[index]),
                             cardFunc: () => showModalBottomSheet(
                               backgroundColor: Colors.white,
                               context: context,
@@ -153,6 +151,11 @@ class _HomeState extends State<Home> {
                                   description: card.description,
                                   weight: card.weight,
                                   price: card.price,
+                                  added: card.added,
+                                  buttonFunc: () {
+                                    changeProductCart(cards[index]);
+                                    Navigator.pop(context);
+                                  }
                                 );
                               },
                             ),
